ConfApplDetailsForm, Applicant Class impplemented
Applicant Class added with a static list which stores the instance as soon as the recruiter fill in the information.
Confirmation Form added so that the recruiter can see how many applicants have been already added with the option to either come back to the dashbord and add another applicant or to go ahead.