Now it is possible to come back to the ConfAppDetailsForm when you wrong clicking on Add; DashForm has a new button which appears only when the list is not empty enabling the previous functionality; When the recruiter logs out, the list and the applicant table are cleaned ; Form3  is now a prototype for the template editor
ConfApplDetailsForm updated so that, when you click add, you have the chance to come back to the previous page;
To fulfill the previous feature, DashForm has a new button which appears only when the applicant list has at least one item. Moreover, when you click logout, the Applicant List is now cleaned as well as the Applicant Table;
New Method in Constant class added which deletes all the values into the Applicant table;
Form3 is now a good prototype for the template editor.