Add codes now on new screen
Tags used to be added on a different form and codes was squeezed onto the code view form. I have made a new screen to add codes while adding tags is now done on the code view screen.
Note that I had to add a databaseDataset object on the dropdown menu in addCode.cs, Otherwise the dropdown would display 'system.data.datarowview' instead of the actual value it was retrieving