ConfApplDetails Form, RegistrationForm and Constants updated so that the system checks whether the user already exists. + Section and Template classes added as well as Editor Class. Template Editor  structure is completed
ConfApplDetails Form, RegistrationForm and Constants updated so that the system checks whether the user already exists.

Now the systems will create templates for all the applicants inserted and it is possible to iterate through them.
Header of the templates get the information as well
Code has been sorted so that there is a division between Back-End and Front-End