ConfApplDetailsForm fixed the value "addApp" with "newApp". Constant Class has a new method which return a query which asks for the sections related to a template;  EditorForm calls two new methods to retrieve sections according to the template selected; LoginForm improved line 111 which checks whether the userPassword.Text is not empty.  Section has both a new method and a new list used to retrieve sections related to a certain template.
ConfApplDetailsForm fixed the value "addApp" with "newApp". Constant Class has a new method which return a query which asks for the sections related to a template;  EditorForm calls two new methods to retrieve sections according to the template selected; LoginForm improved line 111 which checks whether the userPassword.Text is not empty.  Section has both a new method and a new list used to retrieve sections related to a certain template.