Created separate class for editorForm 'Code methods
Editorform is too big. I have removed as much logic to do with inserting, editing and deleting codes and moved it to a new EditorFormCodeClass. code dealing with UI has been left on editorform
An editoformCommon class has been made for logic that is common to codes, sections and templates.